requirements.txt
Django==5.1.1
django-allauth==0.62.1
docx-mailmerge==0.5.0
numpy==1.26.4
pandas==2.2.2
# pywin32==306
requests==2.32.0
XlsxWriter==3.2.0

# for Azure
azure-core==1.32.0

# for deployment

# for local variables
python-dotenv==1.0.1